+++ Runtime/SliderHandleAttribute.cs
@@ -3,18 +3,23 @@ using System.Diagnostics;
 using SaintsField.Interfaces;
 using SaintsField.Utils;
 using UnityEngine;
-// using Debug = UnityEngine.Debug;
 
 namespace SaintsField
 {
     [Conditional("UNITY_EDITOR")]
     [AttributeUsage(AttributeTargets.Field, AllowMultiple = true, Inherited = true)]
-    public class RadiusHandleAttribute: PropertyAttribute, ISaintsAttribute
+    public class SliderHandleAttribute: PropertyAttribute, ISaintsAttribute
     {
         public SaintsAttributeType AttributeType => SaintsAttributeType.Other;
         public string GroupBy => string.Empty;
 
         public readonly string Space;
+
+        public readonly Vector3 Direction;
+        public readonly string DirectionCallback;
+
+        public readonly float Size;
+        public readonly float Snap;
 
         public readonly Vector3 PosOffset;
         public readonly string PosOffsetCallback;
@@ -22,15 +27,25 @@ namespace SaintsField
         public readonly Color Color;
         public readonly string ColorCallback;
 
-        public RadiusHandleAttribute(
+        public SliderHandleAttribute(
             string space = "this",
+            float directionX = 1f, float directionY = 0f, float directionZ = 0f, string directionCallback = null,
+            float size = -1f,
+            float snap = -1f,
             float posXOffset = 0f, float posYOffset = 0f, float posZOffset = 0f, string posOffsetCallback = null,
             EColor eColor = EColor.White, float alpha = 1f, string color = null
         )
         {
             Space = space;
+
+            Direction = new Vector3(directionX, directionY, directionZ);
+            DirectionCallback = RuntimeUtil.ParseCallback(directionCallback).content;
+
+            Size = size;
+            Snap = snap;
+
             PosOffset = new Vector3(posXOffset, posYOffset, posZOffset);
-            PosOffsetCallback = posOffsetCallback;
+            PosOffsetCallback = RuntimeUtil.ParseCallback(posOffsetCallback).content;
 
             Color = eColor.GetColor();
             if (alpha < 1f)