+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 5.18.0 ##
+
+1.  Add: "Runtime Saver" now saves the index of the node for prefab object (instead of plain path) to better find the target object after exit play mode
+2.  Add: "Runtime Saver" now fallback searching the scene's prefabs for the target object if the target can not be found (because of rename, hierarchy changes etc) after exit play mode
+3.  Add: `SliderHandle` scene handle to edit for number / vector types in distance. Useful as a ruler.
+
 ## 5.17.5 ##
 
 Fix: Dependency error [#393](https://github.com/TylerTemp/SaintsField/issues/393)